Add VertexTexturePositionSkinned
