+++ runtime/functor-runtime-common/src/render/mod.rs
@@ -1,4 +1,6 @@
 pub mod vertex;
 mod vertex_position_texture;
+mod vertex_position_texture_skinned;
 
 pub use vertex_position_texture::*;
+pub use vertex_position_texture_skinned::*;
