+++ runtime/functor-runtime-common/src/render/vertex.rs
@@ -11,7 +11,8 @@ pub enum BuiltInVertexChannel {
     Normal,
     Binormal,
     Tangent,
-    BoneIndices,
+    JointIndices,
+    JointWeights,
     // Not bound to any channel used for internal shaders.
     // May be an attribute used for custom shaders
     Custom,
@@ -24,37 +25,7 @@ pub struct VertexAttribute {
     pub size: i32,
 }
 
-#[derive(Debug, Clone)]
-#[repr(C)]
-pub struct VertexPositionTexture {
-    pub position: Vector3<f32>,
-    pub uv: Vector2<f32>,
-}
-
 pub trait Vertex {
     fn get_total_size() -> usize;
     fn get_vertex_attributes() -> Vec<VertexAttribute>;
-}
-
-impl Vertex for VertexPositionTexture {
-    fn get_total_size() -> usize {
-        size_of::<VertexPositionTexture>()
-    }
-
-    fn get_vertex_attributes() -> Vec<VertexAttribute> {
-        vec![
-            VertexAttribute {
-                attribute_channel: BuiltInVertexChannel::Position,
-                attribute_type: VertexAttributeType::Float,
-                offset: offset_of!(VertexPositionTexture, position),
-                size: 3,
-            },
-            VertexAttribute {
-                attribute_channel: BuiltInVertexChannel::Uv,
-                attribute_type: VertexAttributeType::Float,
-                offset: offset_of!(VertexPositionTexture, uv),
-                size: 2,
-            },
-        ]
-    }
 }
